+++ static/js/signup.js
@@ -1,6 +1,6 @@
 $(function () {
 
-    $("#signup-go").on("click", function() {
+    function submitForm() {
         var base = document.getElementById("base-url").getAttribute("href").slice(0, -1);
         var email = $("#signup-email").val();
         var token = $('input[name=csrfmiddlewaretoken]').val();
@@ -18,6 +18,18 @@ $(function () {
         });
 
         return false;
+    }
+
+    $("#signup-go").on("click", submitForm);
+
+    $("#signup-email").keypress(function (e) {
+        if (e.which == 13) {
+            return submitForm();
+        }
     });
+
+    $("#signup-modal").on('shown.bs.modal', function () {
+        $("#signup-email").focus()
+    })
 
 });